backend/requirements.txt
fastapi
uvicorn
python-dotenv
yfinance
huggingface-hub
pandas
requests
